Directory refactor/ create base classes for details MVP
