+++ app/src/main/java/com/derek_s/hubble_gallery/ui/presenters/ActDetailsPresenter.java
@@ -1,0 +1,9 @@
+package com.derek_s.hubble_gallery.ui.presenters;
+
+/**
+ * Created by derek on 15-08-18.
+ */
+
+// private String TAG = getClass().getSimpleName();
+public class ActDetailsPresenter {
+}
