+++ app/src/main/java/com/derek_s/hubble_gallery/ui/views/ActDetailsView.java
@@ -1,0 +1,8 @@
+package com.derek_s.hubble_gallery.ui.views;
+
+/**
+ * Created by derek on 15-08-18.
+ */
+
+public interface ActDetailsView {
+}
